+++ tests/TestCase/Config/GlobalFunctionsTest.php
@@ -25,12 +25,9 @@ namespace MeTools\Test\TestCase;
 use Cake\Filesystem\File;
 use Cake\Filesystem\Folder;
 use Cake\TestSuite\TestCase;
-use MeTools\Core\Plugin;
 
 /**
- * GlobalFunctionsTest class.
- *
- * It tests global functions declared in `config/functions/global.php`.
+ * GlobalFunctionsTest class
  */
 class GlobalFunctionsTest extends TestCase
 {