+++ tests/TestCase/Core/PluginTest.php
@@ -53,10 +53,7 @@ class PluginTest extends TestCase
         $expected = ['MeTools'];
         $this->assertEquals($expected, $result);
 
-        $result = Plugin::load(
-            'TestPlugin',
-            ['path' => 'tests/test_app/Plugin/TestPlugin/src']
-        );
+        $result = Plugin::load('TestPlugin');
         $this->assertNull($result);
 
         $result = Plugin::all();
@@ -67,10 +64,7 @@ class PluginTest extends TestCase
         $expected = ['MeTools'];
         $this->assertEquals($expected, $result);
 
-        $result = Plugin::load(
-            'AnotherTestPlugin',
-            ['path' => 'tests/test_app/Plugin/AnotherTestPlugin/src']
-        );
+        $result = Plugin::load('AnotherTestPlugin');
         $this->assertNull($result);
 
         $result = Plugin::all();
@@ -130,7 +124,7 @@ class PluginTest extends TestCase
         $expected = [ROOT . 'config' . DS . 'bootstrap.php'];
         $this->assertEquals($expected, $result);
 
-        //No one exists
+        //No existing files
         $result = Plugin::path('MeTools', [
             'config' . DS . 'no_existing.php',
             'config' . DS . 'no_existing2.php',