+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.10 branch
+### 2.10.2
+* added `MailHelper` class with `obfuscate()` method;
+* removed `RecaptchaHelper::mailLink()` method;
+* updated for Assets 1.1.0;
+* added tests for `RecaptchaHelper` class;
+* added tests for `FlashComponent` class.
+
 ### 2.10.1
 * fixed serious bug for `BBCodeHelper::remove()` method: now it doesn't remove
     all the HTML code, but only the BBCode code;